feat: NicknameField에 query 적용
#47
